test(eager): CPU compute-device coverage + dtype de-hardcode
Eager boundary saving is pure PyTorch and runs on a CPU compute device,
not just CUDA; add test_cpu_compute_device (acoustic2d / swap2nd) that
builds and differentiates entirely on CPU, guarding against device
hardcoding in the reverse path.  _build/_setup/_first_grad gain a device=
kwarg (default 'cuda', existing call sites unchanged).

Derive ReconState.zero's dtype from the wavefield (compute) dtype instead
of hardcoding float32 — no functional change today (the eager backend is
fp32) but removes a latent assumption in the BS-specific code.

Verified: 27/27 eager BS pytest (incl. the CPU case), 4-way eB/eF cosine
1.0 all equations.

+++ src/sweep/propagator/_eager_boundary_saving.py
@@ -174,7 +174,7 @@ class ReconState:
                  "ring_sizes", "total_cells", "buf", "codes", "scale", "zero")
 
     def __init__(self, nt, rings, shape, n_ring_fields, store_dtype, device,
-                 storage_device=None):
+                 storage_device=None, compute_dtype=torch.float32):
         # ``device`` is the compute device (where the wavefields live); the ring
         # storage lives on ``storage_device`` instead (e.g. host RAM for the 'cpu'
         # offload — trading device memory for per-step H2D/D2H copies).  The zero
@@ -183,7 +183,8 @@ class ReconState:
         self.frame = None
         # Reused read-only zero placeholder for the cpml / unpaired field slots in
         # backward — one buffer for the whole rollout, not a per-step zeros_like.
-        self.zero = torch.zeros(shape, dtype=torch.float32, device=device)
+        # Matches the wavefield (compute) dtype so it composes with the field ops.
+        self.zero = torch.zeros(shape, dtype=compute_dtype, device=device)
         self.rings = rings
         self.checked = False                         # one-time self-check latch
         self.store_dtype = store_dtype
